+++ src/onegov.election_day/onegov/election_day/tests/common.py
@@ -373,6 +373,7 @@ def create_election_compound(client):
     new.form['date'] = date(2015, 1, 1)
     new.form['domain'] = 'canton'
     new.form['elections'] = ['regional-election-a', 'regional-election-b']
+    new.form['party_strengths'] = True
     new.form.submit()
 
 
@@ -403,6 +404,8 @@ def upload_election_compound(client, create=True, canton='gr'):
         new.form['date'] = date(2015, 1, 1)
         new.form['domain'] = 'canton'
         new.form['elections'] = ['regional-election-a', 'regional-election-b']
+        new.form['elections'] = ['regional-election-a', 'regional-election-b']
+        new.form['party_strengths'] = True
         new.form.submit()
 
     for index, slug in enumerate((